Extract task classes from Task: ToDoTask, DeadlineTask, and EventTask
